Increase repetition date based on current day
When user answers flashcard in repetition his repetition data was
changing based on current repetition date. This can be a problem if user
didn't answer flashcard in repetition for example 10. Now user answers
flashcard in repetition and if his next cycle increase is 7 days,
repetition date will be increased for 7 days, which is again behind
current day for 3 days. And that flashcard will now immediately be again
in repetition.