Correct repetition date initial date and TimeCognebusUtils.kt dates
When flashcard is created its repetition date was set with
System.currentTimeMillis() which gives different value than
Calendar.getInstance().timeInMillis and we use Calendar in rest of code.

getBeginningOfCurrentDay would give random value for
Calendar.MILLISECOND(value at its creation point), and when we
getBeginningOfNextDayInMilliseconds (which also gives random
millisecond) to get flashcards for repetition, we would get in situation
that flashcard has almost same repetition date as value we are comparing
it to. In this situation values should be exactly same, and repetition
for flashcard would not be needed, because of < comparator. But here we
get random millisecond values and repetition can be triggered randomly.